Corrected bug that appears when creating multiple subdirs
This bug appeared when it was required to create multiple subdir to host
every files in the source dirs.
Now, everything is fine.
And files with same checksum (basically : duplicate) in the source dir
are not duplicated. (It can be used to identify duplicate using a simple
DB Request in Table "REPOFILES".

Next step :
-> At the end, crypt db and put it in the repository.
-> At the beginning, get the db from the repository and uncrypt it